pyproject.toml
[project]
name = "hitachi-elevator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "pydantic>=2.6.0",
    "click>=8.1.7",
    "uvicorn>=0.27.0",
    "pyyaml>=6.0.1",
    "python-multipart>=0.0.21",
    "openai>=1.0.0",
    "httpx>=0.25.0",
    "elasticsearch>=8.0.0<9.0.0",
    "aiohttp>=3.9.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
]
